feature/unit-ui-testing - separating app and binding module
