+++ app/src/main/java/com/sam43/mindvalleychannels/di/BindingModule.kt
@@ -1,0 +1,41 @@
+package com.sam43.mindvalleychannels.di
+
+import com.sam43.mindvalleychannels.repository.MainRepository
+import com.sam43.mindvalleychannels.repository.MainRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import javax.inject.Inject
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class BindingModule {
+    @Singleton
+    @Binds
+    abstract fun provideCoroutineDispatcher(
+        dispatcherProvider: DefaultDispatcherProvider
+    ): DispatcherProvider
+
+
+    @Singleton
+    @Binds
+    abstract fun provideRepository(
+        mainRepository: MainRepositoryImpl
+    ): MainRepository
+}
+
+interface DispatcherProvider {
+
+    fun main(): CoroutineDispatcher = Dispatchers.Main
+    fun default(): CoroutineDispatcher = Dispatchers.Default
+    fun io(): CoroutineDispatcher = Dispatchers.IO
+    fun unconfined(): CoroutineDispatcher = Dispatchers.Unconfined
+}
+
+class DefaultDispatcherProvider @Inject constructor() :
+    DispatcherProvider
