--- app/src/main/java/com/sam43/mindvalleychannels/di/DispatcherProvider.kt
@@ -1,16 +1,0 @@
-package com.sam43.mindvalleychannels.di
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Inject
-
-interface DispatcherProvider {
-
-    fun main(): CoroutineDispatcher = Dispatchers.Main
-    fun default(): CoroutineDispatcher = Dispatchers.Default
-    fun io(): CoroutineDispatcher = Dispatchers.IO
-    fun unconfined(): CoroutineDispatcher = Dispatchers.Unconfined
-}
-
-class DefaultDispatcherProvider @Inject constructor() :
-    DispatcherProvider
